Backend3/requirements.txt
Flask
flask-pymongo
flask-cors
dnspython
